Add files via upload
This was a program I created to check my home's ip address.  I was being bumped off the network regularly by my service provider, and I wanted to see if my home ip was changing in the process.
Also, I was logging on remotely to my home network via putty, and I wanted to have an up to date ip address available.  The program would send me an email regularly using an hotmail account.  In a later version the ip was recorded by date to an another file, for later analysis.