a water pump control program using Arduino
This was a project to regulate the saltwater level of a fish tank.  A sonic sensor, capable of measuring the height of the water, was set up at the top of a fish tank and used to keep a consistent water level by activating a water pump.

Using an Arduino control board, a sonic sensor was setup above a fish tank and used to measure the water's height.  When the water level dropped below a certain point, the Arduino would send an output signal.  When the water level returned to an acceptable range the output signal would be turned off.  The signal was weak (mA), so an op-amp was built and used to boost the signal to a solenoid.  This solenoid controlled the on-off switch of a water pump.

This code was uploaded into the Arduino and controlled two factors.  First was the activation signal based on the measured water level, averaged over time to control inaccuracies in the sensor.   Second was the cancellation of the signal to prevent overflow of the tank by too much added water.

The design was quite successful, though the salt water in the tank proved corrosive over time, requiring constant replacement of the sensors.